fix: remove duplicate status displays and round ETA seconds in library
Remove redundant status indicators from the actions section - status is now
only shown in the badge area while actions section contains only buttons.
Round ETA seconds to integers to avoid displaying long float values.

+++ resources/views/components/audio-progress-badge.blade.php
@@ -27,7 +27,7 @@
 
     @if($etaSeconds !== null && $etaSeconds > 0)
         <span class="text-zinc-400 dark:text-zinc-500">
-            (~{{ $etaSeconds < 60 ? $etaSeconds . 's' : (int) ceil($etaSeconds / 60) . 'm' }})
+            (~{{ $etaSeconds < 60 ? (int) $etaSeconds . 's' : (int) ceil($etaSeconds / 60) . 'm' }})
         </span>
     @endif
 </span>
